requirements.txt
Bottleneck @ file:///croot/bottleneck_1707864210935/work
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
fonttools @ file:///croot/fonttools_1713551344105/work
importlib-resources @ file:///croot/importlib_resources-suite_1704281845041/work
joblib==1.4.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
matplotlib==3.9.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
packaging @ file:///croot/packaging_1710807400464/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp39-cp39-linux_x86_64.whl#sha256=484b9c988a0aaf85ef9256891a79639c95dfb3cacfd316d465b4f319dfef8d07
pillow @ file:///croot/pillow_1714398848491/work
ply==3.11
psutil==6.0.0
pydicom==2.4.4
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
scikit-learn==1.5.0
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=2cb3d1436f8762a53161486cb9493ccc305fc760e8f830f82888792cbaf99099
seaborn==0.13.2
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.5.0
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tornado @ file:///croot/tornado_1718740109488/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
zipp @ file:///croot/zipp_1704206909481/work
